requirements.txt
albucore==0.0.14
albumentations==1.4.14
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.5
decorator==5.1.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
ftfy==6.2.3
grpcio==1.66.1
h5py==3.11.0
htmlmin==0.1.12
huggingface-hub==0.24.6
idna==3.8
ImageHash==4.3.1
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multimethod==1.12
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pandas-profiling==3.6.6
parso==0.8.4
pathlib==1.0.1
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.1
plotly==5.24.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pyparsing==3.1.4
pyprojroot==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visions==0.7.6
wcwidth==0.2.13
widgetsnbextension==4.0.13
wordcloud==1.9.3
ydata-profiling==4.10.0
zipp==3.20.1
